Refactor DeltaOperationReportImpl
